+ Adding long long to c4, changing rn_stream to uint64.

+++ src/rng/Random_Inline.cc
@@ -8,4 +8,4 @@
 
 #include "Random_Inline.hh"
 
-int rtt_rng::rn_stream;
+uint64_t rtt_rng::rn_stream;
